[IR] Renamed `Backend` to `Ops`.
The API reflection module is called `Ops`, implementations are called

- `LocalOps` (default),
- `FlinkOps` (for Flink),
- `SparkOps` (for Spark).

This is more unified and also avoids name collisions with the
`Compiler` module which is also called `Backend`.
